solution4 updated and added
